backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
python-multipart==0.0.9
numpy==1.26.4
